Introduce AgentBuilderOptions
Sadly we can not defer OltpExport to AgentBuilder.Build() in *all* IHostingEnvironments.

Atleast under WebApplicationHosts the IHostedLifecycleSerivce.StartingAsync happens on

app.Run() at which stage the IServiceCollection is marked as readonly.

To that end we now expose overloads to AddOpenTelemetry taking AgentBuilderOptions.

These are typically only needed for expert level control and not part of the documented getting started
approach.

This PR also addresses all warnings in the solution (branch started off for just that sorry!)