Packaging initial infrastructure (#133)
Not complete but enough to review as part 1 😄 

* Introduces our own version of `OpenTelemetry.AutoInstrumentation`
named `Elastic.OpenTelemetry.AutoInstrumentation`.

Through a bit of msbuild magic we redistribute `instrument.{sh|cmd}` as
`_instrument.{sh|cmd}` and introduce our own `instrument.{sh|cmd}` that
sets the appropiate plugin environment variable before calling the
opentelemetry instrument script.

This allows the invocation to be exactly the same no matter if our
distribution or the vanilla opentelemetry package is installed. A user
can move from or away from Elastic without having to touch orchestration
work. Simply installing the package is enough to move to or away from
us.

* This PR includes the start of a `./build.sh redistribute` build
command that redistributes the opentelemetry autoinstrumentation zip
files that include our AutoInstrumentation plugin.

A follow up of this PR will focus on creating Elastic versions of the
global auto instrumentation installation scripts and ensuring these get
staged as part of the release.