Refactor to use `AddOtlpExporter` and `LoggerProviderBuilder` (#180)
We initially used UseOtlpExporter as a convenient way to enable OTLP
export by default globally for all signals. However, this plays poorly
with the AddOtlpExporter methods, which throw if called by consumer
code. This change avoids that risk. It also introduces a
LoggingProviderBuilder extension to register our defaults, which our
main builder calls.

We still support disabling our default registration of the OTLP exporter
via an environment variable, which the auto instrumentation plugin then
passes when configuring the providers.

This PR also fixes a small typo in the docs.

Closes #177