.NET 9 SDK (#184)
- Updates the solution to use the new .NET 9 SDK. 
- Updates vulnerable dependencies.
- Updates infra and test projects to target .NET 9.
- Updates to Aspire 9 and removes the Aspire workload, which is no
longer required.
- Switches E2E tests to target the QA environment.
- Switches to use 1.9.0 auto instrumentation from OTel.
- Removes `net6.0` targets from projects.
- Cleans up some empty lines in various files.

NOTE: Projects used by auto instrumentation and related tests can't be
updated to use the NET 9 SDK (in the Dockerfile, for example) as the
auto instrumentation library does not yet support .NET 9. That is
scheduled for the 1.10 release later this month.

Once merged, a future PR will update the other dependencies to their
latest versions.