Refactor public API and align with 1.10 and 1.11 of the SDK (#210)
This is quite a significant update from the previous code and includes a
lot of refactorings as a result. In short, as the SDK since 1.10
includes an `OpenTelemetryBuilder`, we can base many extension methods
directly on the interface it implements `IOpenTelemetryBuilder`.

A significant change in this PR is removing the "magic" global override
we used to replace `AddOpenTelemetry` with our behaviour. We no longer
do this to avoid surprises and prefer an explicit small code change from
consumers.

We now also track the calls to the various builders to avoid
bootstrapping components more than is required and log such
circumstances.

We expect the auto instrumentation tests to fail for now as those need
to be uncoupled from the SDK workflow.

Closes #164 
Closes #175