Fixes and refactoring for autoinstrumentation (#232)
Moves common code into `Elastic.OpenTelemetry.Core`. This project is not
packaged, but the code is "compile included" into
`Elastic.OpenTelemetry` and `Elastic.OpenTelemetry.AutoInstrumentation`.
This allows `Elastic.OpenTelemetry.AutoInstrumentation` to contain the
plugin and, with reduced dependencies, avoid any GAC issues.

Also, refactors the `FileLogger` to remove the dependency on
`System.Threading.Channels`, which presented issues when used for auto
instrumentation.