Treat EventLevel.Verbose as debug log level (#288)
When using SDK diagnostics, the lowest supported log level for
OTEL_LOG_LEVEL is 'debug'. To align with this, we emit verbose logs
captured by the EventListener under debug.

This change also updates the logger messages to not use trace level and
prefer debug in those cases. This make diagnostics and support easier as
we only need to instruct the customer to use the levels as supported
upstream.