Update upstream OTel packages (#343)
- Updated package versions in `Directory.Packages.props` to the latest
releases.
- Updated target frameworks in `Packaging.fs` to distribute the .NET 8
TFM for the `net` assets to avoid issues with handler invocation. This
aligns with the TFM of the DLLS distributed for the `net` directory by
the upstream packaging process.
- Updated Dockerfiles to use .NET 10.0 runtime and SDK.
- Refactored `AutoInstrumentationPlugin.cs` to add methods for
configuring resource builders and OTLP exporter options.
- Improved logging messages and corrected date-time format in
`FileLogger`.
- Renamed `TracerProvderBuilderExtensions` to
`ElasticTracerProviderBuilderExtensions` for consistency.
- Improved auto-instrumentation so that we can set user agent for OTLP
exporter
- Fixes some issues with the cross-target compilation of the
`ElasticUserAgentHandler`

General code cleanup and minor refactoring for improved readability and
maintainability.

Closes #327
Closes #338
Closes #191