Enhance logging capabilities (#348)
The following changes were introduced to improve the diagnostics logs to
make it easier for customers to diagnose issues and gain support with
EDOT .NET. We capture more information which is useful for us when
investigating support cases.

- Add more information in the logging premable.
- Add logging of sources for env vars and the final effective values.
- Fix missing debug logs from the OTel SDK events for zero-code use
cases.
- Update date formatting to capture additional resolution.
- Introduce `BootstrapLogger` for detailed logging before main logger
initialization, supporting automatic disposal after inactivity. This is
potentially useful for support scenarios but is marked as experimental.
- Refactor various classes to log public method invocations and
configuration values to the bootstrap logger.
- Add `StackTraceLoggerExtensions` for caller info logging and
`LoadedAssemblyLogHelper` for assembly logging.

Closes #302 
Fixes #309